Refactored types, progressing on the insert function
Refactored the BBoxable type. Now we have a dedicated Node type
for the tree. Each node has a bounding box and items as in children
that are bboxable.

This derailed me from writing the insert method. But I am back at it.
I will try to understand how the depth first search does the insertion.

test: go build